+++ spec/controllers/hyrax/contact_form_controller_spec.rb
@@ -1,0 +1,59 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Hyrax::ContactFormController, type: :controller do
+  let(:user) { User.new(email: 'test@example.com', guest: false) { |u| u.save!(validate: false) } }
+  let(:required_params) do
+    {
+      category: 'Depositing content',
+      name: 'Test Name',
+      email: 'test@test.org',
+      subject: 'Test Subject',
+      message: 'Test Message'
+    }
+  end
+  let(:contact_form) { Hyrax::ContactForm.new(required_params) }
+  let(:params) { required_params }
+
+  routes { Hyrax::Engine.routes }
+  before { sign_in(user) }
+
+  describe '#check_recaptcha' do
+    before do
+      controller.instance_variable_set(:@contact_form, contact_form)
+    end
+
+    context 'when recaptcha is enabled and recaptcha is not verified' do
+      before do
+        allow(controller).to receive(:recaptcha?).and_return(true)
+        allow(controller).to receive(:verify_recaptcha).and_return(false)
+      end
+
+      it 'returns false and throws an error' do
+        expect(controller.check_recaptcha).to eq(false)
+      end
+    end
+
+    context 'when recaptcha is enabled and recaptcha is verified' do
+      before do
+        allow(controller).to receive(:recaptcha?).and_return(true)
+        allow(controller).to receive(:verify_recaptcha).and_return(true)
+      end
+
+      it 'returns a true value' do
+        expect(controller.check_recaptcha).to eq(true)
+      end
+    end
+
+    context 'when recaptcha is not enabled' do
+      before do
+        allow(controller).to receive(:recaptcha?).and_return(false)
+      end
+
+      it 'returns true and processes the email normally' do
+        expect(controller.check_recaptcha).to eq(true)
+      end
+    end
+  end
+end